+++ Sources/NablaScheduling/Views/Scenes/AppointmentConfirmation/AppointmentConfirmationViewController.swift
@@ -56,6 +56,7 @@ final class AppointmentConfirmationViewController: UIViewController {
     
     private lazy var loadingView: UIActivityIndicatorView = {
         let view = UIActivityIndicatorView(style: .large)
+        view.color = NablaTheme.Shared.loadingViewIndicatorTintColor
         view.startAnimating()
         return view
     }()